Updated module augmentation for horizontal-line package.
