+++ packages/ckeditor5-horizontal-line/src/horizontallinecommand.ts
@@ -77,9 +77,3 @@ function getInsertHorizontalLineParent( selection: Selection | DocumentSelection
 
 	return parent as Element;
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		horizontalLine: HorizontalLineCommand;
-	}
-}
