+++ packages/ckeditor5-horizontal-line/src/horizontallineediting.ts
@@ -81,9 +81,3 @@ function toHorizontalLineWidget( viewElement: ViewElement, writer: DowncastWrite
 
 	return toWidget( viewElement, writer, { label } );
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface PluginsMap {
-		[ HorizontalLineEditing.pluginName ]: HorizontalLineEditing;
-	}
-}
